scrapper/pyproject.toml
[tool.poetry]
name = "scrapper"
version = "0.1.0"
description = ""
authors = ["Данилин Даниил <d-danilin@groupprice.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asyncio = "^3.4.3"
httpx = "^0.26.0"
bs4 = "^0.0.2"
uuid = "^1.30"
uvicorn = "^0.27.0.post1"
python-dotenv = "^1.0.1"
fastapi = "^0.109.2"
pydantic = "^2.6.1"
psycopg2-binary = "^2.9.9"

[tool.poetry.scripts]
app = "scrapper.main:loader"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
